does xor and xor in 10000 epochs

+++ test.ps1
@@ -1,4 +1,12 @@
-#rm NNtest.nn
+param(
+    [string]$arg1
+)
+if (!$PSBoundParameters.ContainsKey('arg1')){
+    echo "Please provide the name of the resulting executable."
+    echo "Usage: ./test.ps1 = <main_path>"
+    exit 1
+}
+rm NNtest.nn
 rm log
-./build.ps1 test.exe .\applications\mnist.c
+./build.ps1 test.exe $arg1
 ./test.exe >> log
